modified covariance argument
The argument V to cvBMS.GLM, cvBMS.MGLM, MBI.model and MBI.cvMBI was modified so that, when it is not given or given as None, observations are assumed i.i.d. and all calculations omit V (and its inverse P), simplifying calls to model/cvMBI and making the code more efficient.