added new function
The new MATLAB function "mbicombine" was added to the repository. This is a routine that combines multiple multivariate Bayesian automata (obtained with "mbitrain") and combines their posterior distributions to give a new multivariate Bayesian automaton (suitable for "mbitest"). This function is currently experimental and belongs to a future project on multivariate Bayesian federated learning (MBFL).